requirement.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.15.0
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
crcmod==1.7
cryptography==37.0.2
cycler==0.11.0
Django==4.0.5
django-oss-storage==1.1.1
fonttools==4.33.3
ghostscript==0.7
idna==3.3
jmespath==0.10.0
kiwisolver==1.4.2
matplotlib==3.5.2
mysqlclient==2.1.0
numpy==1.22.4
oss2==2.15.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.9
python-barcode==0.14.0
python-dateutil==2.8.2
pytz==2022.1
qrcode==7.3.1
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
treepoem==3.15.0
tzdata==2022.1
urllib3==1.26.9
